requirements.txt
amqp==2.6.1
asgiref==3.5.2
asttokens==2.2.0
async-timeout==4.0.2
Babel==2.9.0
backcall==0.2.0
billiard==3.6.4.0
bleach==5.0.1
boto==2.49.0
boto3==1.17.75
botocore==1.20.112
celery==4.4.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20230430
dj-database-url==0.5.0
Django==3.2
django-allauth==0.44.0
django-ckeditor==6.0.0
django-countries==7.4.2
django-crispy-forms==1.10.0
django-filter==2.4.0
django-import-export==3.2.0
django-js-asset==2.0.0
django-model-utils==4.1.1
django-phonenumber-field==5.0.0
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-social-share==2.2.1
django-storages==1.11.1
django-summernote==0.8.20.0
django-tz-detect==0.4.0
django-webpush==0.3.4
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
et-xmlfile==1.1.0
executing==1.2.0
gunicorn==20.1.0
http-ece==1.1.0
idna==3.3
install==1.3.5
ipython==8.7.0
jedi==0.18.2
jmespath==0.10.0
jsonfield==3.1.0
kombu==4.6.11
MarkupPy==1.14
matplotlib-inline==0.1.6
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==21.3
parso==0.8.3
paypalrestsdk==1.13.1
pdfkit==1.0.0
pexpect==4.8.0
phonenumbers==8.12.14
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-vapid==1.8.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
pyOpenSSL==23.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
pywebpush==1.9.4
PyYAML==6.0
redis==4.5.4
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.2
swapper==1.1.2.post1
tablib==3.5.0
timezones==2.2.0
traitlets==5.6.0
typing_extensions==4.4.0
urllib3==1.26.9
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
